Added ability to snapshot multiple DBs

+++ tasks/strano_tasks.rake
@@ -25,59 +25,131 @@ namespace :snapshot do
     FileUtils.mkdir_p(snapshot_dir)
     FileUtils.chdir(snapshot_dir)
 
+    STDOUT.puts "\n\n-----[ #{start_time.to_formatted_s(:db)} | CREATING SNAPSHOT ]----------------------------------------------------------------------------------"
+
     # DB
-    STDERR.puts "Dumping #{RAILS_ENV.to_s.upcase} DB to db.sql"
-    config = ActiveRecord::Base.configurations[RAILS_ENV]
-    password = config['password'] ? "-p'#{config['password']}'" : ''
-    mysql_dump_options = [
+    config_keys = []
+    config_dbs = []
+    ActiveRecord::Base.configurations.each do |k,v|
+      config_keys << k
+      config_dbs << v['database']
+    end
+    dbs_to_dump = [ RAILS_ENV ]
 
-        # add-drop-database Add a DROP DATABASE statement before each CREATE DATABASE statement      
-        "--add-drop-database",
+    env_dbs = ENV['DBS']
+    not_env_dbs = ENV['NOT_DBS']
 
-        # add-drop-table Add a DROP TABLE statement before each CREATE TABLE statement
-        "--add-drop-table",
+    env_only_dbs = config_keys.select{ |k| (k =~ %r<#{RAILS_ENV}>) }
 
-        # compact Produce less verbose output
-        "--compact",
+    if env_dbs.present?
+      case env_dbs.downcase
+      when 'all'
+        dbs_to_dump = config_keys
+        # STDERR.puts "  All configs from database.yml: #{dbs_to_dump.join(', ')}"
 
-        #complete-insert Use complete INSERT statements that include column names
-        "--complete-insert",
+      when 'env'
+        dbs_to_dump = config_keys.select{ |k| (k =~ %r<#{RAILS_ENV}>) }
+        # STDERR.puts "  #{RAILS_ENV} only from database.yml: #{dbs_to_dump.join(', ')}"
 
-        # quick Retrieve rows for a table from the server a row at a time
-        "--quick",
+      else
+        STDERR.puts "  Paovided DB list: #{env_dbs}"
+        dbs_to_dump = []
 
-        # --no-data no-data Do not write any table row information (that is, do not dump table contents, only table definitions)
-        # 
-        # --where='where_condition' where Dump only rows selected by the given WHERE condition
-        # --ignore-table=db_name.tbl_name ignore-table Do not dump the given table
-        # 
-        # # DO NOT USE THIS:
-        # --single-transaction single-transaction This option issues a BEGIN SQL statement before dumping data from the server
-        # 
-        # 
-        # 
-        # --no-create-info no-create-info Do not write CREATE TABLE statements that re-create each dumped table      
-        # --no-create-db no-create-db This option suppresses the CREATE DATABASE statements      
-        # --no-autocommit no-autocommit Enclose the INSERT statements for each dumped table within SET autocommit = 0 and COMMIT statements      
-        # --flush-logs flush-logs Flush the MySQL server log files before starting the dump
-        # --disable-keys disable-keys For each table, surround the INSERT statements with disable and enable keys statements
-        # --create-options create-options Include all MySQL-specific table options in the CREATE TABLE statements
-        # 
-        # 
-        # 
-        # tbl_mytable_name –no-data
-    ]
+        env_dbs.split(',').each do |db_name|
+          db_name = RAILS_ENV if db_name == 'env'
+          dbs_to_dump += env_only_dbs.select{ |k| k =~ %r<#{db_name}> }
+          # dbs_to_dump += config_keys.select{ |k| k =~ %r<#{db_name}> }
+        end
+      end
 
-    command_no_password = "mysqldump #{mysql_dump_options.join(' ')} -u #{config['username']} {{password}} #{config['database']} > db.sql"
-    STDERR.puts "Running: #{command_no_password}"
-    `#{command_no_password.gsub("{{password}}", password)}`
+    elsif not_env_dbs.present?
+      initial_dbs = config_keys.select{ |k| (k =~ %r<#{RAILS_ENV}>) }
+      dbs_not_to_dump = []
 
-    # command = `mysqldump #{mysql_dump_options.join(' ')} -u #{config['username']} #{password} #{config['database']} > db.sql`
-    # `#{command}`
-    STDERR.puts "    Done - #{number_to_human_size(File.size("db.sql"))}"
-    STDERR.puts "Compressing db.sql"
-    `tar czf db.tgz db.sql && rm db.sql`
-    STDERR.puts "    Done - #{number_to_human_size(File.size("db.tgz"))}"
+      not_env_dbs.split(',').each do |db_name|
+        dbs_not_to_dump += config_keys.select{ |k| k =~ %r<#{db_name}> }
+      end
+
+      dbs_to_dump = initial_dbs - dbs_not_to_dump
+
+      # STDERR.puts "  Provided NOT_DBS=#{not_env_dbs} - These will be dumped: #{dbs_to_dump.join(', ')}"
+    end
+
+    dbs_to_dump.uniq!
+
+    if dbs_to_dump.blank?
+      STDERR.puts "No matching DBs - defaulting to #{RAILS_ENV}"
+      dbs_to_dump = [ RAILS_ENV ]
+    end
+
+    STDERR.puts "\nDumping DBs: #{dbs_to_dump.join(', ')}\n\n"
+    dbs_to_dump.each do |db_name|
+      rails_env_filename = db_name.gsub(RAILS_ENV, "rails_env")
+      raw_filename = "db_#{rails_env_filename}.sql"
+      tgz_filename = "db_#{rails_env_filename}.tgz"
+      STDERR.puts %Q!
+
+- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+Dumping '#{db_name}' to #{raw_filename}
+
+!
+
+      config = ActiveRecord::Base.configurations[db_name]
+      password = config['password'] ? "-p'#{config['password']}'" : ''
+      mysql_dump_options = [
+
+          # add-drop-database Add a DROP DATABASE statement before each CREATE DATABASE statement      
+          "--add-drop-database",
+
+          # add-drop-table Add a DROP TABLE statement before each CREATE TABLE statement
+          "--add-drop-table",
+
+          # compact Produce less verbose output
+          "--compact",
+
+          #complete-insert Use complete INSERT statements that include column names
+          "--complete-insert",
+
+          # quick Retrieve rows for a table from the server a row at a time
+          "--quick",
+
+          # --no-data no-data Do not write any table row information (that is, do not dump table contents, only table definitions)
+          # 
+          # --where='where_condition' where Dump only rows selected by the given WHERE condition
+          # --ignore-table=db_name.tbl_name ignore-table Do not dump the given table
+          # 
+          # # DO NOT USE THIS:
+          # --single-transaction single-transaction This option issues a BEGIN SQL statement before dumping data from the server
+          # 
+          # 
+          # 
+          # --no-create-info no-create-info Do not write CREATE TABLE statements that re-create each dumped table      
+          # --no-create-db no-create-db This option suppresses the CREATE DATABASE statements      
+          # --no-autocommit no-autocommit Enclose the INSERT statements for each dumped table within SET autocommit = 0 and COMMIT statements      
+          # --flush-logs flush-logs Flush the MySQL server log files before starting the dump
+          # --disable-keys disable-keys For each table, surround the INSERT statements with disable and enable keys statements
+          # --create-options create-options Include all MySQL-specific table options in the CREATE TABLE statements
+          # 
+          # 
+          # 
+          # tbl_mytable_name –no-data
+      ]
+
+      command_no_password = "mysqldump #{mysql_dump_options.join(' ')} -u #{config['username']} {{password}} #{config['database']} > #{raw_filename}"
+      STDERR.puts "  Running: #{command_no_password}"
+      `#{command_no_password.gsub("{{password}}", password)}`
+
+      STDERR.puts "       Raw SQL file size:    #{number_to_human_size(File.size(raw_filename))}"
+      STDERR.puts "   Compressing SQL dump file"
+      `tar czf #{tgz_filename} #{raw_filename} && rm #{raw_filename}`
+      STDERR.puts "       Compressed file size: #{number_to_human_size(File.size(tgz_filename))}\n"
+    end
+
+    STDERR.puts %Q!
+- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+Done dumping DBs
+
+!
 
     # Shared Dirs
     dirs_to_snapshot = []
@@ -96,11 +168,11 @@ namespace :snapshot do
         command = "tar chzf snapshot_dir_#{dir_filename}.tgz -C #{RAILS_ROOT} #{snapshot_directory}"
         STDERR.puts "Compressing #{snapshot_directory}"
         `#{command}`
-        STDERR.puts "    Done - #{number_to_human_size(File.size("snapshot_dir_#{dir_filename}.tgz"))}"
+        STDERR.puts "       Compressed file size: #{number_to_human_size(File.size("snapshot_dir_#{dir_filename}.tgz"))}"
       end
       STDERR.puts "Taring snapshot dirs"
       `tar cf snapshot_dirs.tar snapshot_dir_*.tgz && rm snapshot_dir_*`
-      STDERR.puts "    Done - #{number_to_human_size(File.size('snapshot_dirs.tar'))}"
+      STDERR.puts "       Compressed file size: #{number_to_human_size(File.size("snapshot_dirs.tar"))}"
     end
 
     STDERR.puts "Creating full snapshot"
@@ -111,10 +183,10 @@ namespace :snapshot do
     FileUtils.rm('current.tar') if File.exist?('current.tar')
     snapshot_filename = "#{snapshot_name}.tar"
     `ln -s #{snapshot_filename} current.tar`
-    STDERR.puts "    Done - #{number_to_human_size(File.size(snapshot_filename))}"
+    STDERR.puts "       Compressed file size: #{number_to_human_size(File.size(snapshot_filename))}"
 
     end_time = Time.now
-    STDERR.puts "\nDone - took #{(end_time - start_time).to_f / 60.0} minutes"
+    STDOUT.puts "\n\n-----[ #{end_time.to_formatted_s(:db)} | SNAPSHOT COMPLETE | Took #{((end_time - start_time).to_f / 60.0).round(2)} minutes ]\n"
   end
 
 
@@ -123,6 +195,8 @@ namespace :snapshot do
   task :restore => :environment do
     include ActionView::Helpers::NumberHelper
     start_time = Time.now
+
+    STDOUT.puts "\n\n-----[ #{start_time.to_formatted_s(:db)} | RESTORING SNAPSHOT ]----------------------------------------------------------------------------------\n\n"
 
     snapshots_dir = File.expand_path(File.join(RAILS_ROOT, 'snapshots'))
     FileUtils.mkdir_p(snapshots_dir)
@@ -137,26 +211,47 @@ namespace :snapshot do
     `tar xf #{snapshots_dir}/#{snapshot_file}`
     STDERR.puts "    Done"
 
-    STDERR.puts "Uncompressing db.tgz"
-    `tar xzf db.tgz`
-    STDERR.puts "    Done"
-
-    config = ActiveRecord::Base.configurations[RAILS_ENV]
-    password = config['password'] ? "-p'#{config['password']}'" : ''
-
-    [
-      %Q!DROP DATABASE IF EXISTS #{config['database']}!,
-      %Q!CREATE DATABASE #{config['database']}!,
-    ].each do |mysql_command|
-      `echo "#{mysql_command};" | mysql -u #{config['username']} #{password}`
+    config_keys = []
+    config_dbs = []
+    ActiveRecord::Base.configurations.each do |k,v|
+      config_keys << k
+      config_dbs << v['database']
     end
-    command = %Q[mysql -u #{config['username']} #{password} #{config['database']} < db.sql]
-    STDERR.puts "Restoring db.sql to #{RAILS_ENV} DB"
-    `#{command}`
-    STDERR.puts "    Done"
 
-    FileUtils.rm('db.sql')
-    FileUtils.rm('db.tgz')
+    db_files = Dir["./db_*"]
+    db_files.each do |db_file|
+      if db_file =~ %r<db_(.*)\.tgz>
+        db_name = $1
+        ar_db_name = db_name.gsub("rails_env", RAILS_ENV)
+        raw_filename = "db_#{db_name}.sql"
+        tgz_filename = "db_#{db_name}.tgz"
+
+        if (config = ActiveRecord::Base.configurations[ar_db_name]).blank?
+          STDERR.puts " No entry in database.yml for #{ar_db_name}"
+          next
+        end
+
+        STDERR.puts "Uncompressing #{db_file}"
+        `tar xzf #{db_file}`
+        STDERR.puts "    Done"
+
+        password = config['password'] ? "-p'#{config['password']}'" : ''
+
+        [
+          %Q!DROP DATABASE IF EXISTS #{config['database']}!,
+          %Q!CREATE DATABASE #{config['database']}!,
+        ].each do |mysql_command|
+          `echo "#{mysql_command};" | mysql -u #{config['username']} #{password}`
+        end
+        command = %Q[mysql -u #{config['username']} #{password} #{config['database']} < #{raw_filename}]
+        STDERR.puts "Restoring #{raw_filename} to #{ar_db_name} DB"
+        `#{command}`
+        STDERR.puts "    Done"
+
+        FileUtils.rm(raw_filename)
+        FileUtils.rm(tgz_filename)
+      end
+    end
 
     if File.exist?('snapshot_dirs.tar')
       STDERR.puts "Restoring shared directories"
@@ -198,7 +293,7 @@ namespace :snapshot do
       FileUtils.rm_r(temp_dir)
 
       end_time = Time.now
-      STDERR.puts "\nDone - took %.2f minutes" % ((end_time - start_time).to_f / 60.to_f)
+      STDOUT.puts "\n\n-----[ #{end_time.to_formatted_s(:db)} | RESTORE COMPLETE | Took #{((end_time - start_time).to_f / 60.0).round(2)} minutes ]\n\n"
     end
     
   end
